requirements.txt
selenium
webdriver-manager
beautifulsoup4
lxml
readability-lxml
tiktoken
python-dotenv
requests